% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-rpart.R
\name{augment.rpart}
\alias{augment.rpart}
\title{Augment your model object}
\usage{
\method{augment}{rpart}(x, data = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{rpart model}

\item{data}{data.frame from the model}

\item{newdata}{new data to use for predictions, residuals, etc.}

\item{...}{extra arguments to pass}
}
\value{
\code{augment.rpart} returns the original data with additional columns:
\itemize{
\item \code{.fitted}: The fitted value or class.
\item \code{.resid}: only given when the same data as was used for the model is
provided.
}
}
\description{
Augment your model object
}
\examples{
library(rpart)
rpart_fit <- rpart(Sepal.Width ~ ., iris)
augment(rpart_fit)
}
