% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-gbm.R
\name{tidy.gbm}
\alias{tidy.gbm}
\title{tidy up the model summary of gbm}
\usage{
\method{tidy}{gbm}(x, n_trees = x$n.trees, scale = FALSE,
  sort = TRUE, normalise = TRUE, ...)
}
\arguments{
\item{x}{A \code{gbm} model}

\item{n_trees}{integer. (optional) Number of trees to use for computing
relative importance. Default is the number of trees in x$n.trees. If
not provided, a guess is made using the heuristic: If a test set was
used in fitting, the number of trees resulting in lowest test set error
will be used; else, if cross-validation was performed, the number of
trees resulting in lowest cross-validation error will be used;
otherwise, all trees will be used.}

\item{scale}{(optional) Should importance be scaled? Default is FALSE}

\item{sort}{(optional) Should results be sorted? Default is TRUE}

\item{normalise}{(optional) Should results be normalised to sum to 100? Default is TRUE}

\item{...}{extra functions or arguments}
}
\value{
A tibble containing the importance score for each variable
}
\description{
tidy returns a tibble of variable importance for the rpart pacakge
}
\examples{

# retrieve a tibble of the variable importance from an gbm model

library(gbm)
library(MASS)
fit_gbm <- gbm(calories ~., data = UScereal)

tidy(fit_gbm)

}
