% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-rpart.R
\name{tidy.rpart}
\alias{tidy.rpart}
\title{tidy up the model summary of rpart}
\usage{
tidy.rpart(x, ...)
}
\arguments{
\item{x}{An `rpart` model}

\item{...}{extra functions or arguments}
}
\value{
A tibble containing the importance score for each variable
}
\description{
tidy returns a tibble of variable importance for the rpart pacakge
}
\examples{

# retrieve a tibble of the variable importance from an rpart model

library(rpart)
fit_rpart <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

tidy(fit_rpart)

}
